% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.heading.R
\name{print.heading}
\alias{print.heading}
\title{Heading Printing Function}
\usage{
\method{print}{heading}()
}
\value{
Print out a heading
}
\description{
This function is an auxiliary function for heading printing
}
\author{
Xia Shen
}
